Only set IPM.Note.SMIME on Exchange 2016 and later
* src/cryptcontroller.cpp (CryptController::update_mail_mapi):
Pass exchange major version.
* src/mimemaker.cpp (finalize_message): Add exchange_major_version
param and handle it.

--
This fixes encrypted mails for non exchange servers / older
exchanges. They handle IPM.Note.SMIME apparently not to
spec.

As multipart signed is proven to work but shows the wrong
icon for older exchanges we are conservative and only
use it for exchange major version 15 and later.