Do not show attachment warning when printing
* src/mail.cpp (Mail::parsing_done): Silently remove
unencrypted attachments.

--

When printing we have already shown the warning. So we
should not show it again but silently remove any attachments
that are not hidden before our add_attachments. This
also fixes an issue that when printing sometimes the
child mails which are created for preview and print already
have the decrypted attachments.

GnuPG-Bug-Id: T4560