Restore msg class for S/MIME after move
* src/folder-events.cpp (BeforeMove): Prepare information
needed after move when the mail object is gone and send
the AFTER_MOVE window message.
* src/windmessages.h (AFTER_MOVE): New event.
(wm_after_move_data_t): New.
* src/windowmessages.cpp (AFTER_MOVE): Restore message class
from data.

--
Due to some weird internal stuff in Outlook we have to
change the message class to a custom message class before
move. As the mail object is destoryed to allow the mail
to move we have to load it from the target folder after
the move and then restore it.

This fixes display of S/MIME mails in other MAPI / ActiveSync
S/MIME clients.

GnuPG-Bug-Id: T4543