Improve error handling for large messages
* src/mail.cpp (add_attachments_o): Move into class and
handle error with a message.
* src/mail.h: Update accordingly.
* src/oomhelp.cpp, src/oomhelp.h (add_oom_attachment): Optionally
return the exception description.

--
This removes the "Sorry attachments cannot be removed" warning
by setting m_disable_att_remove_warning during adding because
Outlook sends an AttachmentRemoved event when an Add fails and
additionally shows the error when the Add in OOM fails in
the GUI.

GnuPG-Bug-Id: T4731