Add comment about read only checks / errors
* src/mail.cpp (Mail::preProcessMessage_m): Add a comment.

--
I have tried to check in pre processing if we can modify
a message or not. Strangely, without MAPI_MODIFY access
right we can still modify the OOM body of the mail,
so we can show the decrypted / verified content of the
mail. But we cannot show attachments as adding an
attachment in OOM fails.

So we rely on the error handling for attachment errors.