Improve error handling when adding attachments
* src/mail.cpp (Mail::add_attachments_o): Collect
errors and show a nice message.
* src/oomhelp.cpp (add_oom_attachment): Return
error codes, too.

--
This is mostly important for read only mails where
we do not have the right to add attachments and
with a special case for mails that exceed the maximum
size of the server.

The size limit for us is important because we keep
the original MIME body and temporarily add the
attachments which are part of that body. So
our mails can only have half the size of the
allowed server mail size.