Tune down an error message to a warning
* src/debug.h (log_warn): new Macro.
* src/gpgoladdin.cpp (GpgolAddin::get_instance): Only warn
if instance is gotten when shutting down.

--
This can happen when shutting down Outlook while crypto
messages are open. As part of the closing the message
the instance might be requested to remove the categories.
As we return the instance anyway this is not an error but
ideally we would avoid that by doing it earlier because
the OOM is a bit unstable when we are deep into the
shutdown process.