Improvde sender resolution
* src/mail.cpp (Mail::updateOOMData_o): If both sender and
sendusingaccount can be resolved check if we have have a
key for sender and use that.

--
The logic is that when both sender and sendusingaccount are
found and look like smtp addresses (they don't start with a slash)
then we take the sender address if a key for that address can
be found. Otherwise we use the accounts address as before.

This is important for Institutions that have sender addresses
that differ from their accounts address.

GnuPG-Bug-Id: T4940