Change print detection
* src/application-events.cpp (ItemLoad): Log print.
* src/mail.cpp (Mail::checkIfMailMightBePrinting_o): Use a
global map.

--
This changes the way print jobs are detected by just noting
them down in a set.

Downside of this is that the categories will no longer be
added and decrypted synchronously until the next start of
outlook.

I spent half a day trying to figure out how to better
detect this but I cannot find a way to see if a mail
was actually printed and we get only one BeforePrint
event for all the ItemLoads that occur when you change
the printer as each change of the printer unloads and
reloads the mail.

GnuPG-Bug-Id: T4890