Fix attachment warning for mails from file
* src/oomhelp.cpp (count_visible_attachments): Don't count attachments
without a MAPIOBJECT

--
Testing has shown that for all mail types GpgOL
handles that might contain an unsigned attachment we always
have the MAPIOBJECT / get the hidden state. Only the transient
MIME attachments. The ones used for the MAPI to MIME conversion
and which are hidden by GpgOL and handled by GpgOL will have
no MAPIOBJECT when a mail is opened from file. So this will
remove the warning that "smime.p7m" or "gpgol_mime_structure.txt"
are unsigned and unencrypted attachments.

This also fixes a reference leak for visible attachments
and will probably fix attachment duplication issues.