Add accessor to hidden "Dirty" property
* src/oomhelp.cpp (get_oom_dirty, put_oom_dirty): New.
* src/oomhelp.h: Update accordingly.

--
Stumbled across "You Dirty Rat" by SGriffin. Decided to
play around with this as putting the value might have helped
us preventing outlook from trying to save decrypted data.

Sadly, this didnt work. But it's always good to have
code to work with completely undocumented features. This
property is not even exposed through the introspection
API so its not available in Outlook Spy.