Add oomhelper for folder picker
* src/oomhelp.cpp (oApp): New. Shorthand to get outlook
Application object.
(show_folder_select): Show a folder selection dialog based
on the Outlook object model.

--
This is a first step to evaluate how much effort would
be required to save a decrypted copy of a mail. The
problem is that we can't seem to really have the option
to just write into the target folder but need to move
the decrypted copy.

There is no solution yet how the decrypted mail cannot
be temoporarily stored in a connected MAPI store for
the use case of offline copy.

GnuPG-Bug-Id: T4609