Fix broken encryptSignStart invocation.
* src/mail.cpp (Mail::updateOOMData_o): Set passing the
write for the SaveAs. Advance state to needs actual crypt.
* src/mailitem-events.cpp: Start the encryption explicitly.

--
Commit 0c1bd9076958e584820fadf997ca7d8a248b6888 only worked
because the SaveAs in updateOOMData_o would trigger the
AfterWrite event to do the actual encryption. And while this
worked it is more of a side effect and makes the code
even more spaggethi and not understandable. This now
starts it more explicitly but the whole cryptstate logic
needs more cleanups like this to make it more robust and
finally allow us to get draft encryption working again.