Fix passing keys for encryption
* src/cryptcontroller.cpp (CryptController::lookup_fingerprints):
Do not pop back now that we don't add the last NULL key.

--
Fixes commit da532296d6fbbd3460b4f84e1e511b27e9b2a11f which
checked if a key is null before adding it to the keys.
Previously the last null key was removed with the pop_back. So
now the pop_back removed the last recipient.