Implement BCC splits of Mails
* src/common.h (ERR_SPLIT_UNEXPECTED, ERR_SPLIT_RECIPIENTS): New
bug codes.
* src/cryptcontroller.cpp (CryptController::do_crypto): Do not resolve
again for split copies.
* src/mail.cp (do_crypt): Call dummy pending op abort.
(getKeysForProtocol, Mail::splitCopyMailCallback): Removed. Now
done in RecipientManager
(Mail::splitAndSend_o): Reworked to be iterative.
* src/oomhelp.cpp (set_oom_recipients): Clear out old recipients
when setting new ones.
* src/recipientmanager.cpp (RecipientManager::RecipientManager):
Handle signing keys.
* src/windomessages.cpp (wm_register_pending_op, wm_abort_pending_ops):
New dummies.

--
This finishes the infrastructure required to split of multiple mails
when sending. Error handling with the pending ops and closing
of the mail window on success is still required but in general it
works and two different mails are sent out to different recipient
lists when the BCC Option is set.

GnuPG-Bug-Id: T4814