Clear selection after closing draft
* src/mailitem-events.cpp (CLOSE): Call clear selections after
closing  a draft.
* src/oomhelp.cpp, src/oomhelp.h (oom_clear_selections): New.

--
This avoids problems when a draft is saved and the same mail
is still selected in the explorer. Then it would go out of
sync between the encrypted data and the visible data. To
ensure that the real data is fetched correctly after decyption
we unselect the mail so users have to reselect it.

GnuPG-Bug-Id: T5812