Handle encryption of attachs without MAPI obj
* src/mail.cpp (Mail::add_attachments_o): Store references
to our attachments.
(Mail::updateOOMData_o): Handle MAPIOBJECT null value.

--
This is for the case where we try to encrypt a mail decrypted by
us again. In that case we might not have access to the attachment
data because we cannot access the underlying MAPIOBJECT. Since
we take care not to write it into MAPI. Usually this is not
an issue as the SaveAs call above works. But we can reach this
code in the "BeforeAutoSave" event. In which we cannnot trigger
another write.