Delete temporary enc file if it exists
* src/common.cpp (DeleteFileUtf8): New.
(CreateFileUtf8): SetLastError to avoid having it for conversions.
Do not share delete to avoid problems with attachments.
(get_tmp_outfile): Try to delete existing files.
Fix switch to attachment.dat.
Add error message on failure.

--
This fixes the issues related to empty messages cannot
be encrypted because the temporary file could not be created.

GnuPG-Bug-Id: T5926