Implement simple handling for RFC2231 params
* src/mimedataprovider.cpp (rfc2231_to_2047): Fixup
an RFC2231 string so that it can be parsed by our our rfc2047 parser.
(rfc2231_query_parameter): Add handling for continued parameters
and parameters with encoding / language.

--
For GpgOL only the attachment names are really important
and this should cover most cases without touching the logic
of the overall rfc822 parser. So this should be a fairly
robust solution with little regression risk.

GnuPG-Bug-Id: T6604