Use .mime as the extension for GpgOLs internal attach
* src/mapihelp.h: Use GpgOL_MIME_strucutre.mime instead of .txt

--
These files are visible when the MAPI to MIME conversion has not
happened. E.g. if it has not left your exchange server or is in
your sent mails folder. When you then disable GpgOL or look at
it through the Web Interface you will see these files. Having
a mime file extension will make it easier for us to open them
with Kleopatra.

GnuPG-Bug-Id: T6656