Fix looking at first attachment when body empty
* src/mapihelp.cpp: Check for empty body, too.

--
Due to some weird mail rewrites that happened when filtering
mails we already have the code to check for the content type
of the first attachment to determine if it is a PGP/MIME or
S/MIME mail. But in this case which reportedly came from
gpgtools the body was empty and PR_BODY was not null.

GnuPG-Bug-Id: T6686