Add new option noSaveBeforeDecrypt
* src/addin-options.cpp: Add new strings from config dialog.
* src/common_indep.h (opt): Add noSaveBeforeDecrypt.
* src/mail.cpp (Mail::decryptVerify_o): Do not save the mail
explicitly to MAPI / OOM when noSaveBeforeDecrypt is set.
* src/main.c (read_options): Read new option.

--
This option is by default off and only found under Debug for
potential workarounds. My concern here is that in some corporate
environements which do many modifications on their mail we might
run into some kind of loop that basically causes endless syncs
when a crypto mail is opened. For this reason I would like to
give support the option to disable that feature which we only
built in for a specific problem that we observed at one site.

GnuPG-Bug-Id: T6676