Fix encoding for no singning certificate warning
* src/cryptcontroller.cpp (is_resolved): Use utf8 for gpgol_bug.
Expect the Registry error message to be in local8 bit.

--
This also shows the message every time since it tells the user
to ask the administrator about it and then it does not really
help if the next time the user triggers the action the message
is no longer shown.

GnuPG-Bug-Id: T6683