Fix crash on temporary mailref from ribbon control
* src/ribbon-callbacks.cpp (get_mail_from_control): Check
if the Mail object still exists after releasing the reference.

--
When switching Applications in Outlook and Outlook is configured
to always show a mail in the mailview (e.g. Outlook 2016) it
can happen that the mail is newly read but the invalidation of
the ribbon happens before that and still returns a pointer to
the old mailobject. This will then be deleted once we release
it in the ribbon callback.
This caused a jump into recently freed memory which might
resulted in crashes.

GnuPG-Bug-Id: T6861