Change the GpgOL encryption icon according to its state
* News:Update.
* src/ribbon-callbacks.cpp (get_action_icon)
added new function
* src/ribbon-callbacks.h
declare new get_action_icon function
* src/gpgoladdin.cpp (GpgolRibbonExtender::Invoke)
in switch(dispid) for case IDI_SIGN_ENCRYPT_40PNG call the new
get_action_icon function instead of getIcon
* src/dialogs.h
declare new constances for the 3 new icons and the offset
* src/dialogs.rc
add definition for the 3 new icons
* Addded SVG and PNG version of the 3 icons
* src/icons/enc-only-40.png
* src/icons/enc-only.svg
* src/icons/no-sign-enc-40.png
* src/icons/no-sign-enc.svg
* src/icons/sign-only-40.png
* src/icons/sign-only.svg

--
The new function tries to gets the context from the ribbon control,
then it tries to get the OOM "CurrentItem" from the context as mailitem
and fromit the message via get_oom_base_message. It then uses the
get_gpol_draft_info_flags function on the message to see what action
should be performed on send and calculates the id of the icon to display
based on the returned flags and calls and returns getIcon with the id.

GnuPG-Bug-Id: T7098