Ignore unknown write event again
* src/common_indep.h (opt.closeOnUnknownWriteEvent): New.
* src/mailitem_events.cpp (WriteCancelIgnored): Check for new opt.
* src/main.c (read_options): Read new option.

--
This is an extension of the previouis commit:
10721b1dccf4e89e44ad6a340938acf10f95847f to be even more
conservative when we close our mail since even ignoring
the event only after decryption still leaves the
mark as read issue with encrypted mails if the
marking happens not immediately on read.
To handle situations where this setting might
still be required and for experiments there is
now an option that can be set to enable the closing
of the Mail again. "closeOnUnknownWriteEvent"

GnuPG-Bug-Id: T7223