Fix classification of OpenPGP Multipart/Signed mails
* src/mapihelp.cpp (mapi_change_message_class): Fix
multipart signed handling.
* src/mail.cpp (Mail::decryptVerify_o): Fix class dance
and improve the documentation.

--
For OpenPGP multipart/signed mails the message class set
initially (before the first view) is set to
IPM.Note.SMIME.MultipartSigned. This is why in
change_message_class_ipm_note_smime we check for the
protocol to then set either IPM.Note.GpgOL.MultipartSigned
for PGP mails or IPM.Note.GpgOL.SM.MultipartSigned for SMIME
mails. The old code in mapihelp did not control the flow
properly for that case, which resulted in such mails beeing
classified as IPM.Note.SMIME. Causing T7243. That
bug was introduced in Feburaray 2020 with the WKS handling.

Then in decryptVerify it needs to be adjusted again and is
set back to IPM.Note.SMIME.MultipartSigned since it will
then still display contents and attachments in case GpgOL
is disabled or a client without PGP/MIME Support accesses
it.

GnuPG-Bug-Id: T7242