Revert "Set missing filename to rfc822_email.eml..
This reverts commit 0d16049d41e052ea9b5d7a7878ece3d02a182161.

Since this fix only prints a static filename when an
enacapsulated message is detected but does neither parase the
mail nor fill it with the complete data it is better to
leave it out for a bugfix release as this only creates
more problems then it solves. The encapsulated message part
needs to be eihter completely cut out from our parser and
treated as a MAPI attachment or ignored if we do not support
it.