Change order of overlay dtor
* src/overlay.cpp (Overlay::~Overlay): First enable the
window and then bring it to front before removing the overlay.

--
Since I have seeen hangs here I am unsere if it could be a
problem if the Window related to the overlay is disabled. I
doubt that this will fix an issue but it adds more debug
output and if the overlay does not die properly and
m_overlayCtx->wait(); Hangs at least Outlook will still be
eactive again.