Add handling for titus processed mails
* src/mapihelp.cpp: Add function
get_first_attach_data_with_mime_tag_and_file_name which return the
content of a matching attachment.
The function mapi_get_message_content_type was splitted in 2 with the
new get_content_type_from_header doing the actual processing and the
old map_get_message_type just reading the mail header into a buffer and
calling the new function. The new get_content_type_from_header function
checks if the mail is a multipart/mixed with an X-Titus-* header line
and uses the added function to read the attached mail and calls istself
with these buffer thereby returing the Content-Type of the mail in the
attachment and not the one of the message.